hugging face support, added examples
