+++ src/struct_strm/__about__.py
@@ -1,1 +1,1 @@
-__version__ = "0.0.8"
+__version__ = "0.0.9"
