+++ src/struct_strm/llm_response_handler.py
@@ -1,5 +1,9 @@
 from typing import Union
-from openai.lib.streaming.chat._events import ChunkEvent
+try:
+    from openai.lib.streaming.chat._events import ChunkEvent
+except ImportError:
+    ChunkEvent = None
+
 from dataclasses import field
 
 # List example with openai
@@ -11,6 +15,9 @@ _logger = logging.getLogger(__name__)
 class ContinueSignal(Exception): pass
 
 async def openai_chunk_handler(chunk: ChunkEvent) -> Union[str, ContinueSignal]:
+    if ChunkEvent is None:
+        raise ImportError("OpenAI library is not available. Please install struct-strm with the openai option - pip install struct-strm[openai].")
+    
     if chunk.type == "content.delta":
         chunk = chunk.delta
         _logger.debug(f"Delta: {chunk}")  # get tokens for better mocks
